requirements.txt
# LangChain libraries
langchain
langgraph
langchain-community
langchain-google-vertexai
langchain-ollama
langchain-openai

# Google Cloud AI Platform
google-cloud-aiplatform

# OpenAI library
openai

# aliyun
dashscope

# FAISS library for vector search
faiss-cpu

# Additional libraries
networkx
openpyxl
pandas
pypdf
python-dotenv
rich
pymupdf
# progress bar
tqdm 

# Unstructured library with version constraint
unstructured>=0.6.7